+++ src/cdn-config/util.ts
@@ -13,6 +13,17 @@ export interface ConfigPath {
     env: string
 }
 
+export interface SmartcontractPath {
+    networkId: number,
+    type: string
+}
+
+export const SEASCAPE_CDN = 'https://cdn.seascape.network/';
+
+export let cdnConfigUrl = (path: ConfigPath): string => {
+    return `${SEASCAPE_CDN}${path.project}/${path.env}/config.json`;
+}
+
 export let validateConfNetwork = (networkId: string): boolean => {
     if ((global as any).config === undefined || (global as any).config === null) {
         console.log({