+++ test/read-config.ts
@@ -1,10 +1,9 @@
 import { CdnRead, ConfigPath } from "../src/index";
 
 (async () => {
-    let path = {project: 'lighthouse', env: 'prod'} as ConfigPath;
-    let url = `https://cdn.seascape.network/${path.project}/${path.env}/config.json`;
+    let path = {project: 'lighthouse', env: 'beta'} as ConfigPath;
 
-    let initialized = await CdnRead.initConfig(url, true);
+    let initialized = await CdnRead.initConfig(path, true);
     if (!initialized) {
         console.log(`Global initializiation failed`);
     } else {