+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# CHANGELOG 1.3.3
+* Add the `SmartcontractPath` interface.
+* Hardcode the CDN base URL into the source code and use the utility `ConfigPath` interface to create the URL.
+* The updating the CDN will use the `SmartcontractPath`.
+
 # CHANGELOG 1.3.2
 * Add the optional parameter to CdnRead. If the optional parameter is set to true, then it allows to create an empty global variable.